@@ -74,16 +74,16 @@
     "bazel:test:watch": "ibazel test //..."
   },
   "devDependencies": {
-    "@angular/animations": "6.0.0-beta.7",
-    "@angular/bazel": "6.0.0-beta.7",
-    "@angular/common": "6.0.0-beta.7",
-    "@angular/compiler": "6.0.0-beta.7",
-    "@angular/compiler-cli": "6.0.0-beta.7",
-    "@angular/core": "6.0.0-beta.7",
-    "@angular/http": "6.0.0-beta.7",
-    "@angular/platform-browser": "6.0.0-beta.7",
-    "@angular/platform-browser-dynamic": "6.0.0-beta.7",
-    "@angular/platform-server": "6.0.0-beta.7",
+    "@angular/animations": "6.0.0-rc.1",
+    "@angular/bazel": "6.0.0-rc.1",
+    "@angular/common": "6.0.0-rc.1",
+    "@angular/compiler": "6.0.0-rc.1",
+    "@angular/compiler-cli": "6.0.0-rc.1",
+    "@angular/core": "6.0.0-rc.1",
+    "@angular/http": "6.0.0-rc.1",
+    "@angular/platform-browser": "6.0.0-rc.1",
+    "@angular/platform-browser-dynamic": "6.0.0-rc.1",
+    "@angular/platform-server": "6.0.0-rc.1",
     "@bazel/ibazel": "^0.3.1",
     "@types/express": "^4.0.39",
     "@types/fs-extra": "^4.0.5",
@@ -111,7 +111,7 @@
     "rollup-plugin-node-resolve": "^3.0.2",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "rollup-plugin-uglify": "^2.0.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.0.0-rc.0",
     "sorcery": "^0.10.0",
     "standard-version": "^4.3.0",
     "systemjs": "0.19.43",
@@ -119,7 +119,7 @@
     "tsconfig-paths": "^2.3.0",
     "tslint": "^5.9.1",
     "tsutils": "^2.21.2",
-    "typescript": "~2.6.2",
+    "typescript": "~2.7.2",
     "uglify-js": "^2.8.14",
     "zone.js": "^0.8.12"
   }
