@@ -82,12 +82,8 @@ export const rollupGlobals = {
   '@nguniversal/aspnetcore-engine/tokens': 'nguniversal.aspnetcoreEngine.tokens',
   '@nguniversal/express-engine/tokens': 'nguniversal.expressEngine.tokens',
   '@nguniversal/hapi-engine/tokens': 'nguniversal.hapiEngine.tokens',
-  'rxjs/Observable': 'Rx',
-  'rxjs/operators/filter': 'Rx.operators',
-  'rxjs/operators/map': 'Rx.operators',
-  'rxjs/operators/take': 'Rx.operators',
-  'rxjs/operators/tap': 'Rx.operators',
-  'rxjs/observable/of': 'Rx.Observable',
+  'rxjs': 'Rx',
+  'rxjs/operators': 'Rx.operators',
   'fs': 'fs',
   'express': 'express',
   'hapi': 'hapi',
